backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-openai
langchain-community
langchain-classic
langchain-text-splitters
chromadb
pypdf
